refactor: split vdom helpers into separate files
