+++ src/core/vdom/helpers/normalize-children.js
@@ -1,0 +1,62 @@
+/* @flow */
+
+import { isPrimitive } from 'core/util/index'
+import VNode from 'core/vdom/vnode'
+
+export function normalizeChildren (
+  children: any,
+  ns: string | void,
+  nestedIndex: number | void
+): Array<VNode> | void {
+  if (isPrimitive(children)) {
+    return [createTextVNode(children)]
+  }
+  if (Array.isArray(children)) {
+    const res = []
+    for (let i = 0, l = children.length; i < l; i++) {
+      const c = children[i]
+      const last = res[res.length - 1]
+      //  nested
+      if (Array.isArray(c)) {
+        res.push.apply(res, normalizeChildren(c, ns, i))
+      } else if (isPrimitive(c)) {
+        if (last && last.text) {
+          last.text += String(c)
+        } else if (c !== '') {
+          // convert primitive to vnode
+          res.push(createTextVNode(c))
+        }
+      } else if (c instanceof VNode) {
+        if (c.text && last && last.text) {
+          last.text += c.text
+        } else {
+          // inherit parent namespace
+          if (ns) {
+            applyNS(c, ns)
+          }
+          // default key for nested array children (likely generated by v-for)
+          if (c.tag && c.key == null && nestedIndex != null) {
+            c.key = `__vlist_${nestedIndex}_${i}__`
+          }
+          res.push(c)
+        }
+      }
+    }
+    return res
+  }
+}
+
+function createTextVNode (val) {
+  return new VNode(undefined, undefined, undefined, String(val))
+}
+
+function applyNS (vnode, ns) {
+  if (vnode.tag && !vnode.ns) {
+    vnode.ns = ns
+    if (vnode.children) {
+      for (let i = 0, l = vnode.children.length; i < l; i++) {
+        applyNS(vnode.children[i], ns)
+      }
+    }
+  }
+}
